Create API crud blog, brand

+++ server/service/productCategoryService.js
@@ -60,11 +60,26 @@ const updateProductCategoryService = asyncHandler(async (req, res) => {
 
   const { pcid } = req.params
 
+  // Get the schema keys of the Blog model
+  const schemaKeys = Object.keys(ProductCategory.schema.paths)
+
+  // Check if the keys in req.body are in the schema keys
+  for (let key in req.body) {
+    if (!schemaKeys.includes(key)) {
+      return {
+        error: true,
+        errorReason: `Key "${key}" does not exist for updating`,
+        success: false,
+        object: `Key "${key}" không tồn tại để cập nhật`
+      }
+    }
+  }
+
   const response = await ProductCategory.findByIdAndUpdate(pcid, req.body, { new: true })
 
   return {
     error: response ? false : true,
-    errorReason: response ? 'Update product categoriy successfully' : 'Update product categoriy failed',
+    errorReason: response ? 'Update product category successfully' : 'Update product category failed',
     success: response ? true : false,
     object: response ? response : 'Update product category failed'
   }
@@ -74,11 +89,24 @@ const deleteProductCategoryService = asyncHandler(async (req, res) => {
 
   const { pcid } = req.params
 
+  // Get the schema keys of the Product Category model
+  const schemaKeys = Object.keys(ProductCategory.schema.paths)
+
+  // Check if bid exists in the schema keys
+  if (!schemaKeys.includes('pcid')) {
+    return {
+      error: true,
+      errorReason: `Invalid product category ID: ${pcid}`,
+      success: false,
+      object: `ID nhóm sản phẩm không hợp lệ: ${pcid}`
+    }
+  }
+
   const response = await ProductCategory.findByIdAndDelete(pcid)
 
   return {
     error: response ? false : true,
-    errorReason: response ? 'Delete product categoriy successfully' : 'Delete product categoriy failed',
+    errorReason: response ? 'Delete product category successfully' : 'Delete product category failed',
     success: response ? true : false,
     object: response ? response : 'Delete product category failed'
   }
